+++ src/realm.ts
@@ -2,27 +2,10 @@ import { faker } from '@faker-js/faker';
 import * as t from 'io-ts';
 import { RealmStorage } from './realm-storage';
 import { isMappedRef, ManifestedRef, MappedRef } from './ref';
-import {
-  Define,
-  DefineTraversal,
-  EntityC,
-  Manifest,
-  Persist,
-  Traversal,
-} from './types';
-
-const identityTraversal: Traversal<unknown> = {
-  is: (_value: unknown): _value is unknown => true,
-  traverse: f => x => f(x),
-};
+import { Define, EntityC, Manifest, Persist } from './types';
 
 export class Realm {
   private readonly storage = new RealmStorage();
-  private readonly traversals: Traversal<any>[] = [identityTraversal];
-
-  readonly defineTraversal: DefineTraversal = traversal => {
-    this.traversals.push(traversal);
-  };
 
   readonly define: Define = (
     Entity,
@@ -75,9 +58,19 @@ export class Realm {
       return manifestedRef.mappedValue;
     };
 
-    const maybeProcessRef = (value: unknown) => {
+    const processValue = (value: unknown) => {
       if (isMappedRef(value)) {
         return processRef(value);
+      }
+
+      if (t.UnknownRecord.is(value)) {
+        const acc: Record<string, unknown> = {};
+
+        for (const [subKey, subValue] of Object.entries(value)) {
+          acc[subKey] = processValue(subValue);
+        }
+
+        return acc;
       }
 
       return value;
@@ -88,11 +81,7 @@ export class Realm {
         continue;
       }
 
-      for (const traversal of this.traversals) {
-        if (traversal.is(value)) {
-          manifestedEntity[key] = traversal.traverse(maybeProcessRef)(value);
-        }
-      }
+      manifestedEntity[key] = processValue(value);
     }
 
     for (const key in overrides) {