+++ src/ref.test.ts
@@ -4,17 +4,12 @@ import * as t from 'io-ts';
 import { either, option } from 'io-ts-types';
 import { Realm } from './realm';
 import { Ref } from './ref';
-import { ManifesterOptions, Traversal } from './types';
+import { ManifesterOptions } from './types';
 
 describe('Ref', () => {
   describe('when used inside of an `Option`', () => {
     it(`calls the referenced entity's manifester`, () => {
       const realm = new Realm();
-
-      realm.defineTraversal({
-        is: option(t.unknown).is,
-        traverse: O.map,
-      });
 
       const Author = t.strict({ id: t.string });
 
@@ -49,16 +44,9 @@ describe('Ref', () => {
   });
 
   describe('when used inside of an `Either`', () => {
-    const eitherTraversal: Traversal<E.Either<unknown, unknown>> = {
-      is: either(t.unknown, t.unknown).is,
-      traverse: f => E.bimap(f, f),
-    };
-
     describe('inside of a `Left`', () => {
       it(`calls the referenced entity's manifester`, () => {
         const realm = new Realm();
-
-        realm.defineTraversal(eitherTraversal);
 
         const Author = t.strict({ id: t.string });
 
@@ -96,8 +84,6 @@ describe('Ref', () => {
       it(`calls the referenced entity's manifester`, () => {
         const realm = new Realm();
 
-        realm.defineTraversal(eitherTraversal);
-
         const Author = t.strict({ id: t.string });
 
         const Post = t.strict({
@@ -128,6 +114,80 @@ describe('Ref', () => {
 
         expect(authorManifester).toHaveBeenCalled();
       });
+    });
+  });
+
+  describe('when used inside of a deeply-nested object', () => {
+    it(`calls the referenced entity's manifester`, () => {
+      const realm = new Realm();
+
+      const Author = t.strict({ id: t.string });
+
+      const Post = t.strict({
+        id: t.string,
+        a: t.strict({
+          u: t.strict({
+            t: t.strict({
+              h: t.strict({
+                o: t.strict({
+                  r: t.strict({
+                    id: t.string,
+                  }),
+                }),
+              }),
+            }),
+          }),
+        }),
+      });
+
+      const authorManifester = jest.fn<
+        t.TypeOf<typeof Author>,
+        [ManifesterOptions]
+      >(({ faker }) => Author.encode({ id: faker.datatype.uuid() }));
+
+      realm.define(Author, {
+        manifest: authorManifester,
+      });
+
+      realm.define(Post, {
+        manifest: ({ faker }) =>
+          Post.encode({
+            id: faker.datatype.uuid(),
+            a: {
+              u: {
+                t: {
+                  h: {
+                    o: {
+                      r: {
+                        id: Ref.to(Author).through(author => author.id),
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          }),
+      });
+
+      const post = realm.manifest(Post);
+
+      expect(post).toMatchObject({
+        a: {
+          u: {
+            t: {
+              h: {
+                o: {
+                  r: {
+                    id: expect.any(String),
+                  },
+                },
+              },
+            },
+          },
+        },
+      });
+
+      expect(authorManifester).toHaveBeenCalled();
     });
   });
 });